BigQuery/requirements.txt
astroid==2.11.6
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
db-dtypes==1.0.2
dill==0.3.5.1
flatten-json==0.1.13
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
platformdirs==2.5.2
proto-plus==1.20.6
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.14.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tomli==2.0.1
tomlkit==0.11.0
typed-ast==1.5.4
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
